% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_body}
\alias{clean_address_body}
\title{Clean address entry bodies}
\usage{
clean_address_body(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean body of address entries provided.
}
\examples{
clean_address_body(c("1S20 Londun st.", "13<J Dixon rd"))
}
