% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_ends}
\alias{clean_address_ends}
\title{Clean ends in address entries}
\usage{
clean_address_ends(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean ends in provided address entries.
}
\examples{
clean_address_ends(c(" -; 18, 20 London st.", ",,12 Dixon st.$  "))
}
