% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_mac}
\alias{clean_address_mac}
\title{Standardise "Mac" prefix in address entries}
\usage{
clean_address_mac(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to standardise "Mac" prefix in provided address entries.
}
\examples{
\dontrun{
  clean_address_mac(c("McWilliam", "M'Donald", "M c Fyfe"))
}
}
