% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_names}
\alias{clean_address_names}
\title{Clean place names in address entries}
\usage{
clean_address_names(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean place names in provided address entries.
}
\examples{
clean_address_names(c("18, 20 Londun Street", "136 Dixn Road"))
}
