% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_number}
\alias{clean_address_number}
\title{Clean address entry numbers}
\usage{
clean_address_number(address)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean number of address entries provided.
}
\examples{
clean_address_number(c("1S20 Londun st.", "13<J Dixon rd"))
}
