% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_others}
\alias{clean_address_others}
\title{Miscellaneous cleaning operations in address entries}
\usage{
clean_address_others(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Carries out miscellaneous cleaning operations in provided address entries.
}
\examples{
\dontrun{
  clean_address_others(c("18, 20 London. Street, Agents", "136 Dixon Road d"))
}
}
