% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_places}
\alias{clean_address_places}
\title{Clean places in address entries}
\usage{
clean_address_places(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean places in provided address entries: street, road, place,
quay, etc.
}
\examples{
clean_address_places(c("Bothwell Cir.", "Railway arch."))
}
