% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_possessives}
\alias{clean_address_possessives}
\title{Standardise possessives in address entries}
\usage{
clean_address_possessives(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to standardise possessives in provided address entries.
}
\examples{
clean_address_possessives(c("Adam's Court", "Aird's Lane"))
}
