% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_post_clean}
\alias{clean_address_post_clean}
\title{Post-cleaning operation for address entries}
\usage{
clean_address_post_clean(address)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Performs post-cleaning operations on provided address entries.
}
\examples{
clean_address_post_clean(c(": 1820 London Street.   -", "136 Saint Enoch Square,;"))
}
