% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_pre_clean}
\alias{clean_address_pre_clean}
\title{Pre-cleaning operation for address entries}
\usage{
clean_address_pre_clean(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Performs pre-cleaning operations on provided address entries.
}
\examples{
\dontrun{
  clean_address_pre_clean(c(": 1S20 Londun st.   -", "13<J st enoch sq,;"))
}
}
