% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_saints}
\alias{clean_address_saints}
\title{Clean "Saint" prefix in address entries}
\usage{
clean_address_saints(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean "Saint" prefix in provided address entries.
}
\examples{
clean_address_saints(c("St Georges st.", "St Enoch sq"))
}
