% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_suffixes}
\alias{clean_address_suffixes}
\title{Clean unwanted suffixes in address entries}
\usage{
clean_address_suffixes(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean unwanted suffixes in provided address entries.
}
\examples{
\dontrun{
  clean_address_suffixes("36 Rose street so")
}
}
