% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_worksites}
\alias{clean_address_worksites}
\title{Clean worksites in address entries}
\usage{
clean_address_worksites(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean worksites in provided address entries.
}
\examples{
clean_address_worksites("Woodyd")
}
