% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename}
\alias{clean_forename}
\title{Clean entries forename}
\usage{
clean_forename(names)
}
\arguments{
\item{names}{A character string vector of forename(s).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean provided forename.
}
\section{Details}{

Single letter forenames are standardised to the forename starting with that
letter occurring the most frequently in the dataset. i.e A. -> Alexander,
B. -> Bernard, C. -> Colin, D. -> David, etc.
}

\examples{
clean_forename(c("McWm.", "Jas."))
}
