% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_punctuation}
\alias{clean_forename_punctuation}
\title{Standardise punctuation in forenames}
\usage{
clean_forename_punctuation(names)
}
\arguments{
\item{names}{A character string vector of names.}
}
\value{
A vector of character strings.
}
\description{
Attempts to standardise punctuation in provided name entries.
}
\examples{
clean_forename_punctuation(c("William;", "Miss", "John,Hugh"))
}
