% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_separate_words}
\alias{clean_forename_separate_words}
\title{Separate double-barrelled forenames}
\usage{
clean_forename_separate_words(names)
}
\arguments{
\item{names}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to separate double-barrelled forenames in provided name entries.
}
\examples{
clean_forename_separate_words(c("William", "Alexander", "JohnHugh"))
}
