% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_spelling}
\alias{clean_forename_spelling}
\title{Clean forenames spelling}
\usage{
clean_forename_spelling(names)
}
\arguments{
\item{names}{A character string vector of names.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean spelling in provided name entries.
}
\examples{
clean_forename_spelling(c("Wm.", "A.", "Jn Huh"))
}
