% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_mac}
\alias{clean_mac}
\title{Standardise "Mac" prefix in people's name}
\usage{
clean_mac(names)
}
\arguments{
\item{names}{A character string vector of names.}
}
\value{
A vector of character strings.
}
\description{
Attempts to standardise "Mac" prefix in provided name entries.
}
\examples{
\dontrun{
  clean_mac("McWilliam")
}
}
