% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_name_ends}
\alias{clean_name_ends}
\title{Clean ends in entries names}
\usage{
clean_name_ends(names)
}
\arguments{
\item{names}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean ends in provided name entries.
}
\examples{
clean_name_ends(c("William-", "$Alexander", "John Hugh   "))
}
