% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_occupation}
\alias{clean_occupation}
\title{Clean entries occupation}
\usage{
clean_occupation(occupations)
}
\arguments{
\item{occupations}{A character string vector of occupation(s).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean provided occupation
}
\examples{
\dontrun{
  clean_occupation(c("Wine and spirit mercht", "Bkr"))
}
}
