% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_parentheses}
\alias{clean_parentheses}
\title{Clean entries of in brackets information}
\usage{
clean_parentheses(x)
}
\arguments{
\item{x}{A character string vector.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean entries of unwanted information displayed in brackets.
}
\examples{
clean_parentheses("Miller street. (Agents).")
}
