% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_specials}
\alias{clean_specials}
\title{Clean entries special characters}
\usage{
clean_specials(x)
}
\arguments{
\item{x}{A character string vector.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean entries of unwanted special characters
}
\examples{
\dontrun{
  clean_specials(c("Wine and spirit \u00bbmerchant", "Mac\u00bb William"))
}
}
