% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_string_ends}
\alias{clean_string_ends}
\title{Clean string ends}
\usage{
clean_string_ends(strings)
}
\arguments{
\item{strings}{A character string vector.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean ends of strings provided
}
\examples{
\dontrun{
  clean_string_ends(c(" -; 18, 20 Mary hill.*", ",,12 Dixon Street^"))
}
}
