% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_surname_punctuation}
\alias{clean_surname_punctuation}
\title{Standardise punctuation in forenames}
\usage{
clean_surname_punctuation(names)
}
\arguments{
\item{names}{A character string vector of names.}
}
\value{
A vector of character strings.
}
\description{
Attempts to standardise punctuation in provided name entries.
}
\examples{
\dontrun{
  clean_surname_punctuation(c("Abbott*", "*Abercromby", "Blair"))
}
}
