% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_title}
\alias{clean_title}
\title{Clean entries name title}
\usage{
clean_title(titles)
}
\arguments{
\item{addresses}{A character string vector of title(s).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean titles attached to names provided: Captain, Major, etc.
}
\examples{
clean_title(c("  Capt Wm. ABOT", "?-Rev Alex ABRCROMBIE"))
}
