% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_get_address_house_type}
\alias{combine_get_address_house_type}
\title{Get house address column type}
\usage{
combine_get_address_house_type(column)
}
\arguments{
\item{column}{A Character string: ends in "house.number" or "house.body".}
}
\value{
A Character string: "number" or "body".
}
\description{
Identifies the type of the house address column provided: number or body.
}
