% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_has_match_failed}
\alias{combine_has_match_failed}
\title{Check for failed matches}
\usage{
combine_has_match_failed(number, body)
}
\arguments{
\item{number}{A vector of address numbers. Integer or character string.}

\item{body}{A character string vector of address bodies.}
}
\value{
A boolean vector.
}
\description{
Provided with two equal length vectors, returns TRUE for indexes where both
entries are "NA" and FALSE otherwise.
}
\examples{
\dontrun{
  numbers <- c("18, 20", NA)
  bodies <- c("London Road.", NA)
  combine_has_match_failed(numbers, bodies)
}
}
