% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_label_failed_matches}
\alias{combine_label_failed_matches}
\title{Label failed matches}
\usage{
combine_label_failed_matches(directory)
}
\arguments{
\item{directory}{A Scottish post office directory in the form of a data.frame
or other object that inherits from the data.frame class such as a
\code{\link[tibble]{tibble}}. Columns must at least include
\code{address.house.number}, \code{address.house.body}.}
}
\value{
A data.frame of the same class as the one provided in \code{directory}.
Columns include \code{address.house.number}, \code{address.house.body}. For entries
for which both \code{address.house.number} and \code{address.house.body} are \code{NA},
\code{address.house.number} and \code{address.house.body} are labelled as "" and
"Failed to match with general directory" respectively.
}
\description{
Labels failed matches as such in the provided Scottish post office directory
data.frame.
}
