% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_label_failled_matches}
\alias{combine_label_failled_matches}
\title{Label failed matches}
\usage{
combine_label_failled_matches(directory)
}
\arguments{
\item{directory}{A directory dataframe. Columns must include \code{address.house.number},
\code{address.house.body}.}
}
\value{
A dataframe.
}
\description{
Labels failed matches as such in the provided directory dataframe.
}
\examples{
directory <- data.frame(
  page = c("71", "71"),
  surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
  occupation = c("Wine and spirit merchant", "Baker"),
  address.trade.number = c("18, 20", "12"),
  address.house.number = c("136", NA),
  address.trade.body = c("London Road", "Dixon Place"),
  address.house.body = c("Queen Square", NA),
  stringsAsFactors = FALSE
)
combine_label_failled_matches(directory)
}
