% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_label_if_match_failed}
\alias{combine_label_if_match_failed}
\title{Label failed matches}
\usage{
combine_label_if_match_failed(type = c("number", "body"), ...)
}
\arguments{
\item{type}{A Character string: "number", "body". Type of column to label.}

\item{...}{Further arguments to be passed down to
\code{\link{combine_has_match_failed}}}
}
\value{
A character string vector.
}
\description{
Labels failed matches as such.
}
\examples{
\dontrun{
  numbers <- c("18, 20", NA)
  bodies <- c("London Road.", NA)
  combine_label_if_match_failed("number", number = numbers, body = bodies)
  combine_label_if_match_failed("body", number = numbers, body = bodies)
}
}
