% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_make_match_string}
\alias{combine_make_match_string}
\title{Mutate operation(s) in directory dataframe trade address}
\usage{
combine_make_match_string(directory)
}
\arguments{
\item{directory}{A directory dataframe. Columns must include \code{forename},
\code{surname}, \code{occupation}, \code{address.trade.number}, \code{address.trade.body}.}
}
\value{
A dataframe.
}
\description{
Creates a 'match.string' column in the provided directory dataframe composed
of entries full name and trade address pasted together.
Missing trade address entries are replaced with a random generated string.
}
\section{Details}{

The purpose of the 'match.string' column is to facilitates the matching of
the general to trades directory down the line. It allows to calculate a
string distance metric between each pair of entries and match those falling
below a specified threshold.
}

\examples{
directory <- data.frame(
  page = c("71", "71"),
  surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
  occupation = c("Wine and spirit merchant", "Baker"),
  address.trade = c("18, 20, London Road.", "No trade address found"),
  stringsAsFactors = FALSE
)
combine_make_match_string(directory)
}
\seealso{
\code{\link{match_records}} for the matching of the general to
trades directory.
}
