% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_match_general_to_trades}
\alias{combine_match_general_to_trades}
\title{Match general to trades directory records}
\usage{
combine_match_general_to_trades(
  trades_directory,
  general_directory,
  progress = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{trades_directory}{A trades directory dataframe. Columns must include
\code{surname}, \code{forename}, \code{address.trade.number}, \code{address.trade.body}.}

\item{general_directory}{A general directory dataframe. Columns must include
\code{surname}, \code{forename}, \code{address.trade.number}, \code{address.trade.body},
\code{address.house.number}, \code{address.house.body}.}

\item{progress}{Whether progress should be shown (\code{TRUE}) or not (\code{FALSE}).}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}

\item{...}{Further arguments to be passed down to
\code{\link[fuzzyjoin]{stringdist_left_join}}.}
}
\value{
A tibble
}
\description{
Attempts to complement trades directory dataframe with house address
information from the general directory dataframe provided by matching records
from the two datasets using the distance metric specified.
}
\examples{
trades_directory <- tibble::tibble(
  page = rep("71", 3L),
  rank = c("135", "326", "586"),
  surname = c("Abbott", "Abercromby", "Blair"),
  forename = c("William", "Alexander", "John Hugh"),
  occupation = c("Wine and spirit merchant", "Baker", "Victualler"),
  type = rep("OWN ACCOUNT", 3L),
  address.trade.number = c("18, 20", "12", "280"),
  address.trade.body = c("London Road", "Dixon Place", "High Street")
)
general_directory <- tibble::tibble(
  page = rep("71", 2L),
  surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
  occupation = c("Wine and spirit merchant", "Baker"),
  address.trade.number = c("18, 20", ""),
  address.house.number = c("136", "29"),
  address.trade.body = c("London Road", "Dixon Place"),
  address.house.body = c("Queen Square", "Anderston Quay")
)
combine_match_general_to_trades(
 trades_directory, general_directory, progress = TRUE, verbose = FALSE,
 method = "osa", max_dist = 5
)

}
