% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_match_general_to_trades_progress}
\alias{combine_match_general_to_trades_progress}
\title{Match general to trades directory records}
\usage{
combine_match_general_to_trades_progress(
  trades_directory,
  general_directory,
  verbose,
  ...
)
}
\arguments{
\item{trades_directory}{A trades directory dataframe. Columns must include
\code{surname}, \code{forename}, \code{address.trade.number}, \code{address.trade.body}.}

\item{general_directory}{A general directory dataframe. Columns must include
\code{surname}, \code{forename}, \code{address.trade.number}, \code{address.trade.body},
\code{address.house.number}, \code{address.house.body}.}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}

\item{...}{Further arguments to be passed down to
\code{\link[fuzzyjoin]{stringdist_left_join}}.}
}
\value{
A dataframe.
}
\description{
Attempts to complement trades directory dataframe with house address
information from the general directory dataframe provided by matching records
from the two datasets using the distance metric specified. Shows a progress
bar indicating function progression.
}
\examples{
trades_directory <- data.frame(
  page = c("71", "71"),
  surname = c("Abbott", "Abercromby", "Blair"),
  forename = c("William", "Alexander", "John Hugh"),
  occupation = c("Wine and spirit merchant", "Baker", "Victualler"),
  address.trade.number = c("18, 20", "12", "280"),
  address.trade.body = c("London Road", "Dixon Place", "High Street"),
  stringsAsFactors = FALSE
)
general_directory <- data.frame(
  page = c("71", "71"),
  surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
  occupation = c("Wine and spirit merchant", "Baker"),
  address.trade.number = c("18, 20", ""),
  address.house.number = c("136", "29"),
  address.trade.body = c("London Road", "Dixon Place"),
  address.house.body = c("Queen Square", "Anderston Quay"),
  stringsAsFactors = FALSE
)
combine_match_general_to_trades_plain(
 trades_directory, general_directory, verbose = FALSE,
 method = "osa", max_dist = 5
)
}
\seealso{
\code{\link{combine_match_general_to_trades}}.
}
