% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_no_trade_address_to_random_string}
\alias{combine_no_trade_address_to_random_string}
\title{Mutate operation(s) in directory data.frame address.trade column.}
\usage{
combine_no_trade_address_to_random_string(directory)
}
\arguments{
\item{directory}{A Scottish post office directory in the form of a data.frame
or other object that inherits from the data.frame class such as a
\code{\link[tibble]{tibble}}. Columns must at least include \code{address.trade}.}
}
\value{
A data.frame of the same class as the one provided in \code{directory};
columns include at least \code{address.trade}.
}
\description{
Replaces missing trade address(es) in the provided Scottish post office
directory data.frame with random string(s). Random string(s) only show(s)
in body of trade address entry(/ies).
}
\section{Details}{

Prevents unwarranted matches when matching general to trades directory.
Unrelated records with similar name and trade address entry labelled as
missing would be otherwise matched.
}

