% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_no_trade_address_to_randon_string}
\alias{combine_no_trade_address_to_randon_string}
\title{Mutate operation(s) in directory dataframe trade address}
\usage{
combine_no_trade_address_to_randon_string(directory)
}
\arguments{
\item{directory}{A directory dataframe. Columns must include \code{address.trade},}
}
\value{
A dataframe.
}
\description{
Replaces missing trade addresses in the provided directory dataframe with
random string. Random string only shows in body of trade address entries.
}
\section{Details}{

Prevents unwarranted matches when matching general to trades directory.
Unrelated records with similar name and trade address entry labelled as
missing would be otherwise matched.
}

\examples{
\dontrun{
  directory <- data.frame(
    page = rep("71", 2L),
    rank = c("135", "326"),
    surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
    occupation = c("Wine and spirit merchant", "Baker"),
    type = rep("OWN ACCOUNT", 3L),
    address.trade = c("18, 20, London Road.", "No trade address found"),
    stringsAsFactors = FALSE
  )
  combine_no_trade_address_to_randon_string(directory)
}
}
