% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_random_string_if_no_address}
\alias{combine_random_string_if_no_address}
\title{Conditionally return a random string}
\usage{
combine_random_string_if_no_address(address)
}
\arguments{
\item{address}{A character string..}
}
\value{
A length 1 character string vector.
}
\description{
Returns a 22 character long random string if address provided is labeled as
missing.
}
\examples{
combine_random_string_if_no_address(c("18, 20 London Road", "No trade address found"))
}
