% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_clean_directory}
\alias{general_clean_directory}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_clean_directory(directory, progress = TRUE, verbose = FALSE)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{occupation},
\code{forename}, \code{surname} and \code{addresses}.}

\item{progress}{Whether progress should be shown (\code{TRUE}) or not (\code{FALSE}).}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}
}
\value{
A tibble
}
\description{
Attempts to clean the provided general directory dataframe provided.
}
\examples{
directory <- tibble::tibble(
  page = rep("71", 2L),
  surname = c("ABOT", "ABRCROMBIE"), forename = c("Wm.", "Alex"),
  occupation = c("Wine and spirit mercht — See Advertisement in Appendix.", ""),
  addresses = c(
    "1S20 Londn rd; ho. 13<J Queun sq",
    "Bkr; I2 Dixon Street, & 29 Auderstn Qu.; res 2G5 Argul st."
  )
)
general_clean_directory(directory, progress = TRUE, verbose = FALSE)

}
