% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_clean_directory_progress}
\alias{general_clean_directory_progress}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_clean_directory_progress(directory, verbose)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{occupation},
\code{forename}, \code{surname} and \code{addresses}.}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}
}
\value{
A dataframe.
}
\description{
Attempts to clean the provided general directory dataframe provided. Shows a
progress bar indication the progression of the function.
}
\examples{
directory <- data.frame(
  page = c("71", "71"),
  surname = c("ABOT", "ABRCROMBIE"), forename = c("Wm.", "Alex"),
  occupation = c("Wine and spirit mercht — See Advertisement in Appendix.", ""),
  addresses = c(
    "1S20 Londn rd; ho. 13<J Queun sq",
    "Bkr; I2 Dixon Street, & 29 Auderstn Qu.; res 2G5 Argul st."
  ),
  stringsAsFactors = FALSE
)
general_clean_entries(directory, verbose = FALSE)
}
