% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_fix_structure}
\alias{general_fix_structure}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_fix_structure(directory, verbose)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{addresses}.}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}
}
\value{
A dataframe.
}
\description{
Attempts to fix the structure of the raw general directory dataframe provided.
For each entry, attempts to fix parsing errors by moving pieces if information
provided in the right columns, attempts to separate trade from house address,
attempt to separate multiple trade addresses, attempt to separate number from
address body.
}
\examples{
\dontrun{
  directory <- data.frame(
    page = c("71", "71"),
    surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
    occupation = c("Wine and spirit merchant; house", ""),
    addresses = c(
      "18, 20 London Street",
      "Baker; 12 Dixon Street, & 29 Anderston Quay; and 265 Argyle Street."
    ),
    stringsAsFactors = FALSE
  )
  general_fix_structure(directory, verbose = FALSE)
}
}
