% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_move_house_to_address}
\alias{general_move_house_to_address}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_move_house_to_address(directory, regex)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{occupation}
and  \code{addresses}.}

\item{regex}{Regex to use for the task provided as a character string.}
}
\value{
A dataframe.
}
\description{
For some raw general directory entries 'house' referring to address type lives
in the occupation column as a result of parsing errors
\code{general_move_house_to_address} attempts to move this information in the
appropriate destination, the \code{addresses} column.
}
\examples{
directory <- data.frame(
  page = c("71", "71", "71"),
  surname = c("ABOT", "ABRCROMBIE", "BLAI"), forename = c("Wm.", "Alex", "Jn Huh"),
  occupation = c(
    "Wine and spirit merchant; house", "Baker", "Victualer"
  ),
  addresses = c("1820 Mary hill", "", "280, High stret"),
  stringsAsFactors = FALSE
)
regex <- globals_regex_house_to_address
general_move_house_to_address(directory, regex)
}
