% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_repatriate_occupation_from_address}
\alias{general_repatriate_occupation_from_address}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_repatriate_occupation_from_address(directory, regex)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{occupation} and
\code{addresses}.}

\item{regex}{Regex to use for the task provided as a character string.}
}
\value{
A dataframe.
}
\description{
For some raw general directory entries occupation information lives in the
addresses column as a result of parsing errors
\code{general_repatriate_occupation_from_address}
attempts to move this information in the appropriate destination, the
\code{occupation} column.
}
\examples{
\dontrun{
  directory <- data.frame(
    page = c("71", "71", "71"),
    surname = c("ABOT", "ABRCROMBIE", "BLAI"), forename = c("Wm.", "Alex", "Jn Huh"),
    occupation = c(
      "", "Wine and spirit merchant", ""
    ),
    addresses = c("bkr; 1820, Mary hill", "", "Victualer; 280, High stret"),
    stringsAsFactors = FALSE
  )
  regex <- globals_regex_occupation_from_address
  general_repatriate_occupation_from_address(directory, regex)
}
}
