% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_split_address_numbers_bodies}
\alias{general_split_address_numbers_bodies}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_split_address_numbers_bodies(
  directory,
  regex_split_address_numbers,
  regex_split_address_body,
  regex_split_address_empty,
  ignore_case_filter,
  ignore_case_match
)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include
\code{address.trade} and \code{address.house}.}

\item{regex_split_address_numbers}{Regex to use to match address number(s).}

\item{regex_split_address_body}{Regex to use to match address body(/ies).}

\item{regex_split_address_empty}{Regex to use to match empty address entries.}

\item{ignore_case_filter}{Boolean specifying whether case should be ignored
(\code{TRUE}) or not (\code{FALSE}) for using one of the regexes above as filtering
regex in \code{\link{utils_regmatches_if_found}}.}

\item{ignore_case_match}{Boolean specifying whether case should be ignored
(\code{TRUE}) or not (\code{FALSE}) for using one of the regexes above as matching
regex in \code{\link{utils_regmatches_if_found}}.}
}
\value{
A dataframe.
}
\description{
Attempts to separate number from body of address entries in general directory
dataframe provided.
}
\examples{
\dontrun{
  directory <- data.frame(
    page = c("71", "71"),
    surname = c("ABOT", "ABRCROMBIE"), forename = c("Wm.", "Alex"),
    occupation = c("Wine and spirit merchant", "Baker"),
    address.trade = c("18, 20 London Street", "12 Dixon Street"),
    address.house = c("136 Queen Street", "265 Argyle Street"),
    stringsAsFactors = FALSE
  )
  regex_split_address_numbers <- globals_regex_split_address_numbers
  regex_split_address_body <- globals_regex_split_address_body
  regex_split_address_empty <- globals_regex_split_address_empty
  general_split_address_numbers_bodies(
    directory, regex_split_address_numbers, regex_split_address_body,
    regex_split_address_empty, ignore_case_filter = TRUE, ignore_case_match = TRUE
  )
}
}
