% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_split_trade_addresses}
\alias{general_split_trade_addresses}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_split_trade_addresses(
  directory,
  regex_split,
  ignore_case_split,
  regex_filter,
  ignore_case_filter,
  regex_match,
  ignore_case_match
)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{addresses.trade}.}

\item{regex_split}{Regex to use to split addresses.}

\item{ignore_case_split}{Boolean specifying whether case should be ignored
(\code{TRUE}) or not (\code{FALSE}) for \code{regex_split} above.}

\item{regex_filter}{Regex to use to search for address entries with post-split
undesired leftovers.}

\item{ignore_case_filter}{Boolean specifying whether case should be ignored
(\code{TRUE}) or not (\code{FALSE}) for \code{regex_filter} above.}

\item{regex_match}{Regex to use to clear address entries from post-split
undesired leftovers.}

\item{ignore_case_match}{Boolean specifying whether case should be ignored
(\code{TRUE}) or not (\code{FALSE}) for \code{regex_match} above.}
}
\value{
A dataframe.
}
\description{
Attempts to separate multiple trade addresses in the general directory dataframe
provided for entries for which more than one are provided.
}
\examples{
\dontrun{
  directory <- data.frame(
    page = c("71", "71"),
    surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
    occupation = c("Wine and spirit merchant", "Baker"),
    addresses.trade = c(
      "18, 20 London Street, 136 Queen Street; 134, 136 South Portland Street",
      "12 Dixon Street, & 29 Anderston Quay; and 265 Argyle Street."
    ),
    stringsAsFactors = FALSE
  )
  regex_split <- globals_regex_split_trade_addresses
  regex_filter <- globals_regex_and_filter
  regex_match <- globals_regex_and_match
  general_split_trade_addresses(
    directory, regex_split, regex_filter, regex_match,
    ignore_case_split = FALSE, ignore_case_filter = TRUE, ignore_case_match = FALSE
  )
}
}
