% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{general_split_trade_house_addresses}
\alias{general_split_trade_house_addresses}
\title{Mutate operation(s) in general directory dataframe column(s)}
\usage{
general_split_trade_house_addresses(directory, regex, verbose)
}
\arguments{
\item{directory}{A general directory dataframe. Columns must include \code{addresses}.}

\item{regex}{Regex to use for the task provided as a character string.}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}
}
\value{
A dataframe.
}
\description{
Attempts to separate trade and house addresses in the general directory dataframe
provided for entries for which both are provided.
}
\examples{
directory <- data.frame(
  page = c("71", "71"),
  surname = c("Abbott", "Abercromby"), forename = c("William", "Alexander"),
  occupation = c("Wine and spirit merchant", "Baker"),
  addresses = c(
    "18, 20 London Street; ho. 136 Queen Street.",
    "12 Dixon Street; res, 29 Anderston Quay."
  ),
  stringsAsFactors = FALSE
)
regex <- globals_regex_house_split_trade
general_split_trade_house_addresses(directory, regex, verbose = FALSE)
}
