% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_ampersand}
\alias{globals_ampersand}
\title{Ampersand in directory entries}
\format{
A data frame with 3 variables:
\describe{
\item{pattern}{regex for ampersand reading error matching}
\item{replacement}{replacement pattern for substitution operations}
\item{ignore_case}{
boolean operator indicating whether the corresponding
regex is case sensitive or not.
}
}
}
\usage{
globals_ampersand
}
\description{
A dataset containing regular expression meant to match common (OCR) errors in
reading the ampersand character: "&" in directory entries. For each error
pattern a replacement pattern is provided for used in substitution operations
as well as a boolean operator indicating whether the corresponding regex is
case sensitive or not.
}
\keyword{datasets}
