% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_ampersand_vector}
\alias{globals_ampersand_vector}
\title{Ampersand in directory entries}
\format{
A character string vector.
}
\usage{
globals_ampersand_vector
}
\description{
A character vector of regular expressions to match common (OCR) errors in
reading the ampersand character: "&" in directory entries.
}
\keyword{datasets}
