% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_and_double_quote}
\alias{globals_and_double_quote}
\title{Ampersand in directory entries}
\format{
A character string vector.
}
\usage{
globals_and_double_quote
}
\description{
A character vector of regular expressions to match common (OCR) errors in
reading the ampersand character: "&" in directory entries.
}
\section{Details}{

Some regexes contain the double quote character: '"'.
}

\keyword{datasets}
