% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_general_colnames}
\alias{globals_general_colnames}
\title{General directory column names}
\format{
A character string vector.
}
\usage{
globals_general_colnames
}
\description{
A character vector of column names for general directories.
}
\keyword{datasets}
