% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_macs}
\alias{globals_macs}
\title{"Mac" pre-fixes in name entries}
\format{
A data frame with 3 variables:
\describe{
\item{pattern}{regex for "Mac" pre-fix matching}
\item{replacement}{replacement pattern for substitution operations}
\item{ignore_case}{
boolean operator indicating whether the corresponding
regex is case sensitive or not.
}
}
}
\usage{
globals_macs
}
\description{
A dataset containing regular expression meant to match commonly (OCR) misread
"Mac" pre-fixes in directory name entries. For each "Mac" pre-fix a replacement
pattern is provided for used in substitution operations as well as a boolean
operator indicating whether the corresponding regex is case sensitive or not.
}
\keyword{datasets}
