% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_places_raw}
\alias{globals_places_raw}
\title{Place types in address entries}
\format{
A character string vector.
}
\usage{
globals_places_raw
}
\description{
A character vector of common place types found in directory address entries
}
\keyword{datasets}
