% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_address_prefix}
\alias{globals_regex_address_prefix}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_address_prefix
}
\description{
Regular expression used to remove undesired pre-fixes in general
directory address records.
}
\seealso{
\code{\link{utils_remove_address_prefix}}
}
\keyword{datasets}
