% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_and_filter}
\alias{globals_regex_and_filter}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_and_filter
}
\description{
Regular expression used to the word "and" in a filtering operation part of a
mutate operation in the general directory provided.
}
\seealso{
\code{\link{general_split_trade_addresses}}
}
\keyword{datasets}
