% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_get_address_house_type}
\alias{globals_regex_get_address_house_type}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_get_address_house_type
}
\description{
Regular expression used in the making of the match.string that eventually
enables the matching of general and trades directory records.
}
\seealso{
\code{\link{combine_get_address_house_type}}
}
\keyword{datasets}
