% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_house_split_trade}
\alias{globals_regex_house_split_trade}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_house_split_trade
}
\description{
Regular expression used to separate trades from house addresses in general
directory.
}
\seealso{
\code{\link{general_split_trade_house_addresses}}
}
\keyword{datasets}
