% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_house_to_address}
\alias{globals_regex_house_to_address}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_house_to_address
}
\description{
Regular expression used to move the word "house" from the occupation
column to the addresses column in general directory.
}
\seealso{
\code{\link{general_move_house_to_address}}
}
\keyword{datasets}
