% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_irrelevants}
\alias{globals_regex_irrelevants}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_irrelevants
}
\description{
Regular expression used to match irrelevant information in the directory
dataset provided.
}
\seealso{
\code{\link{utils_clear_irrelevants}}
}
\keyword{datasets}
