% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_occupation_from_address}
\alias{globals_regex_occupation_from_address}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_occupation_from_address
}
\description{
Regular expression used to repatriate occupation from address column in general
directory.
}
\seealso{
\code{\link{general_repatriate_occupation_from_address}}
}
\keyword{datasets}
