% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_split_address_empty}
\alias{globals_regex_split_address_empty}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_split_address_empty
}
\description{
Regular expression used to separate numbers from body in provided general
directory address entries.
}
\seealso{
\code{\link{general_split_address_numbers_bodies}}
}
\keyword{datasets}
