% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_split_trade_addresses}
\alias{globals_regex_split_trade_addresses}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_split_trade_addresses
}
\description{
Regular expression used to split multiple trade addresses when more than one
are provided.
}
\seealso{
\code{\link{utils_remove_address_prefix}}
}
\keyword{datasets}
