% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_regex_titles}
\alias{globals_regex_titles}
\title{Regular expression for mutate operations in directory datasets}
\format{
A character string vector.
}
\usage{
globals_regex_titles
}
\description{
Regular expression used to match title in provided directory name entries.
}
\keyword{datasets}
