% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_suffixes}
\alias{globals_suffixes}
\title{Address suffixes}
\format{
A data frame with 3 variables:
\describe{
\item{pattern}{regex for suffix matching}
\item{replacement}{replacement pattern for substitution operations}
\item{ignore_case}{
boolean operator indicating whether the corresponding
regex is case sensitive or not.
}
}
}
\usage{
globals_suffixes
}
\description{
A dataset containing regular expression meant to match commonly (OCR) misread
suffixes in directory address entries. For each suffix a replacement
pattern is provided for used in substitution operations as well as a boolean
operator indicating whether the corresponding regex is case sensitive or not.
}
\keyword{datasets}
