% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_surnames}
\alias{globals_surnames}
\title{Surnames in directory records}
\format{
A data frame with 3 variables:
\describe{
\item{pattern}{regex for surname matching}
\item{replacement}{replacement pattern for substitution operations}
\item{ignore_case}{
boolean operator indicating whether the corresponding
regex is case sensitive or not.
}
}
}
\usage{
globals_surnames
}
\description{
A dataset containing regular expression meant to match commonly (OCR) misread
surnames in directory name entries. For each surname a replacement
pattern is provided for used in substitution operations as well as a boolean
operator indicating whether the corresponding regex is case sensitive or not.
}
\keyword{datasets}
