% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_trades_colnames}
\alias{globals_trades_colnames}
\title{Trades directory column names}
\format{
A character string vector.
}
\usage{
globals_trades_colnames
}
\description{
A character vector of column names for trades directories.
}
\keyword{datasets}
