% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.r
\docType{data}
\name{globals_union_colnames}
\alias{globals_union_colnames}
\title{Combined directories column names}
\format{
A character string vector.
}
\usage{
globals_union_colnames
}
\description{
A character vector of column names for the dataset where general directory
records are matched to trades directory records.
}
\keyword{datasets}
