% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_clean_address_body}
\alias{utils_clean_address_body}
\title{Clean address(es) body}
\usage{
utils_clean_address_body(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A vector of character strings.
}
\description{
Clean body record of provided address(es).
}
\examples{
\dontrun{
  utils_clean_address_body(
    c("London st.", "Mary hill.*", "&;Dixon st.", "Craigrownie, Cove.$")
  )
}
}
