% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_clean_address_ends}
\alias{utils_clean_address_ends}
\title{Clean address entry ends}
\usage{
utils_clean_address_ends(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A vector of character strings.
}
\description{
Clean beginning and end of the provided address entries.
}
\examples{
\dontrun{
  utils_clean_address_ends(
    c(
      " -; 18, 20 London st.; house, Mary hill.*",
      ",,12 &;Dixon st.; residence, Craigrownie, Cove.$"
    )
  )
}
}
