% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_clean_address_number}
\alias{utils_clean_address_number}
\title{Clean address(es) number}
\usage{
utils_clean_address_number(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A vector of character strings.
}
\description{
Clean number record of provided address(es).
}
\examples{
\dontrun{
  utils_clean_address_number(c(" -; 1820", ",,12"))
}
}
