% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_clean_ends}
\alias{utils_clean_ends}
\title{Clean entry ends}
\usage{
utils_clean_ends(directory, ...)
}
\arguments{
\item{directory}{A directory dataframe.}

\item{...}{Columns to clean provided as expressions.}
}
\value{
A dataframe.
}
\description{
Clean entry ends for the specified columns in the directory dataframe
provided
}
\examples{
\dontrun{
  directory <- data.frame(
    page = c("71", "71", "71"),
    surname = c("ABOT", "ABRCROMBIE", "BLAI"), forename = c("Wm.", "Alex", "Jn Huh"),
    occupation = c("Wine and spirit merchant", "Baker", "Victualer"),
    address.trade.number = c(" -; 1820", "", "280"),
    address.trade.body = c("London st. ; house, Mary hill.*", "", "High stret"),
    stringsAsFactors = FALSE
  )
  utils_clean_ends(directory, address.trade.number, address.trade.body)
}
}
