% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_clear_content}
\alias{utils_clear_content}
\title{Clear string of matched content}
\usage{
utils_clear_content(string_search, regex_content, ignore_case)
}
\arguments{
\item{string_search}{Character string to search for match(es).}

\item{regex_content}{PCRE type regex provided as a character string
of match(es) to search for.}

\item{ignore_case}{Boolean specifying whether case should be ignored (\code{TRUE})
or not (\code{FALSE}).}
}
\value{
A character string.
}
\description{
Clears the provided string of the content specified as a regex.
}
\examples{
\dontrun{
  utils_clear_content("glasgow-entrepreneurs", "^.+-", TRUE)
}
}
