% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_execute}
\alias{utils_execute}
\title{Execute function}
\usage{
utils_execute(verbose, fun, ...)
}
\arguments{
\item{verbose}{Boolean specifying whether to silence the function execution
(\code{FALSE}) or not (\code{TRUE}).}

\item{fun}{Function to execute provided as an expression.}

\item{...}{Argument(s) to be passed to the function above for execution.}
}
\value{
Whatever the provided function returns.
}
\description{
Executes the function provided. Execution can be silenced via the \code{verbose}
parameter.
}
\examples{
\dontrun{
  utils_execute(TRUE, message, "I'm showing in console")
}
}
