% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_IO_load}
\alias{utils_IO_load}
\title{Load object into memory}
\usage{
utils_IO_load(...)
}
\arguments{
\item{...}{Destination parameters to be passed to \code{\link{utils_IO_path}}.}
}
\value{
R object from destination \code{.rds} file.
}
\description{
Load saved object as \code{.rds} file back into memory.
}
\examples{
\dontrun{
  utils_IO_load("home/projects", "glasgow-entrepreneurs")
}
}
