% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_IO_path}
\alias{utils_IO_path}
\title{Make path for input/output operations}
\usage{
utils_IO_path(directory_path, ..., extension)
}
\arguments{
\item{directory_path}{Path to directory where \code{file_name} lives as
character string.}

\item{...}{File name components provided as character strings to be passed
down to \code{\link{utils_make_file}}.}

\item{extension}{File extension as character string}
}
\value{
Path to destination file as a character string.
}
\description{
Paste provided path to directory and file name provided using '/' as
separator.
}
\examples{
\dontrun{
  utils_IO_path("home/projects", "glasgow-entrepreneurs", "csv")
}
}
