% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_IO_write}
\alias{utils_IO_write}
\title{Write object to long term memory}
\usage{
utils_IO_write(data, ...)
}
\arguments{
\item{data}{R object to save.}

\item{...}{Destination parameters to be passed to \code{\link{utils_IO_path}}.}
}
\value{
No return value, called for side effects.
}
\description{
Save the object provided to specified path as \code{.rds} file.
}
\examples{
\dontrun{
  utils_IO_write(mtcars, "home/projects", "mtcars")
}

}
