% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_is_address_missing}
\alias{utils_is_address_missing}
\title{Check is address entry not missing}
\usage{
utils_is_address_missing(type)
}
\arguments{
\item{type}{A character string: "house" or "trade", specifying the type of
address to check.}
}
\value{
A Boolean vector: TRUE if both number and body are empty.
}
\description{
Checks whether or not for each address in the evaluation environment, body
and number are filled/not empty.
}
\section{Details}{

The function is for primarily use in the
\code{\link{utils_label_address_if_missing}} function called by
\code{\link{utils_label_missing_addresses}} where it provides a filtering
vector used for labelling missing addresses. \code{utils_is_address_missing} creates
an expression and further evaluates it two levels up in the environment tree,
in other words in the directory dataframe eventually passed down to
\code{\link{utils_label_missing_addresses}}.
}

