% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_label_address_if_missing}
\alias{utils_label_address_if_missing}
\title{Label addresses if missing}
\usage{
utils_label_address_if_missing()
}
\value{
A character string vector of address bodies, unchanged if provided,
labelled as missing otherwise.
}
\description{
If address is empty label body accordingly: "no house/trade address
found".
}
\section{Details}{

The function is for primarily use in the
\code{\link{utils_label_missing_addresses}} function where it provides a
vector of address bodies \code{utils_label_address_if_missing} creates an
expression and further evaluates it one level up in the environment tree,
in other words in the directory dataframe eventually passed down to
\code{\link{utils_label_missing_addresses}}.
}

