% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_make_file}
\alias{utils_make_file}
\title{Make file name}
\usage{
utils_make_file(..., extension)
}
\arguments{
\item{...}{File name component(s) as character string(s).}

\item{extension}{File extension as character string}
}
\value{
File name as a character string.
}
\description{
Pastes the arguments provided together using '-'. Appends result string with
the extension provided.
}
\examples{
utils_make_file("glasgow", "entrepreneurs", extension = "csv")

}
