% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_make_path}
\alias{utils_make_path}
\title{Make destination path}
\usage{
utils_make_path(...)
}
\arguments{
\item{...}{Path components as character string(s).}
}
\value{
Path to last element provided as a character string.
}
\description{
Pastes the arguments provided together using '/' as separator.
}
\examples{
utils_make_path("home", "projects", "glasgow-entrepreneurs.csv")

}
