% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_mute}
\alias{utils_mute}
\title{Mute a function call execution}
\usage{
utils_mute(fun, ...)
}
\arguments{
\item{fun}{Function to execute as an expression.}

\item{...}{Argument(s) to be passed to the function above for execution.}
}
\value{
Whatever the provided function in \code{fun} returns.
}
\description{
Executes the function provided while silencing the potential messages related
to its execution
}
\examples{
\dontrun{
  utils_mute(message, "I'm not showing in console")
}
}
