% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_split_and_name}
\alias{utils_split_and_name}
\title{Split string into tibble}
\usage{
utils_split_and_name(string, pattern, num_col, colnames)
}
\arguments{
\item{string}{Character string to be split.}

\item{pattern}{Pattern to split on as character string (can be a regex).}

\item{num_col}{Number of parts to split the string into as integer.}

\item{colnames}{Column names for the output tibble.}
}
\value{
A \code{\link[tibble]{tibble}}
}
\description{
Split provided string according to specified pattern. Organise output as a
\code{\link[tibble]{tibble}}.
}
\examples{
\dontrun{
  utils_split_and_name("glasgow-entrepreneurs", "-", 2, c("location", "occupation"))
}
}
