% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{utils_squish_all_columns}
\alias{utils_squish_all_columns}
\title{Clear extra white spaces in dataframe}
\usage{
utils_squish_all_columns(df)

utils_squish_all_columns(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.

A dataframe.
}
\description{
Removes blanks (white spaces and tabs) at the beginning and end of all entries
of the provided dataframe. Converts all series of white space and/or tab(s)
in the body of all dataframe entries into a single white space.

Removes blanks (white spaces and tabs) at the beginning and end of all entries
of the provided dataframe. Converts all series of white space and/or tab(s)
in the body of all dataframe entries into a single white space.
}
\examples{
\dontrun{
  df <- data.frame(
    location = "  glasgow ", occupation = "wine    merchant",
    stringsAsFactors = FALSE
  )
  df <- utils_squish_all_columns(df)
}
\dontrun{
  df <- data.frame(
    location = "  glasgow ", occupation = "wine    merchant",
    stringsAsFactors = FALSE
  )
  df <- utils_squish_all_columns(df)
}
}
