% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_attached_words}
\alias{clean_address_attached_words}
\title{Clean attached words in address entries}
\usage{
clean_address_attached_words(addresses)
}
\arguments{
\item{addresses}{A character string vector of addresses.}
}
\value{
A vector of character strings.
}
\description{
Attempts to separate attached words in provided address entries.
}
\examples{
\dontrun{
  clean_address_attached_words(c("18, 20 LondonSt.", "12 Dixon.st."))
}
}
