% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_attached_words}
\alias{clean_address_attached_words}
\title{Clean attached words in address entry(/ies)}
\usage{
clean_address_attached_words(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) cleaned of attached words.
}
\description{
Attempts to separate attached words in provided address entry(/ies).
}
