% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_body}
\alias{clean_address_body}
\title{Clean address entry(/ies) body}
\usage{
clean_address_body(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with cleaned bodies.
}
\description{
Attempts to clean body of address entry(/ies) provided.
}
