% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_ends}
\alias{clean_address_ends}
\title{Clean ends in address entry(/ies)}
\usage{
clean_address_ends(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with clean ends.
}
\description{
Attempts to clean ends in provided address entry(/ies).
}
