% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_names}
\alias{clean_address_names}
\title{Clean place name(s) in address entry(/ies)}
\usage{
clean_address_names(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with clean name(s).
}
\description{
Attempts to clean place names in provided address entry(/ies).
}
