% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_number}
\alias{clean_address_number}
\title{Clean address entry numbers}
\usage{
clean_address_number(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with cleaned numbers.
}
\description{
Attempts to clean number of address entry(/ies) provided.
}
