% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_others}
\alias{clean_address_others}
\title{Miscellaneous cleaning operations in address entry(/ies)}
\usage{
clean_address_others(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of clean address(es).
}
\description{
Carries out miscellaneous cleaning operations in provided address entry(/ies).
}
