% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_possessives}
\alias{clean_address_possessives}
\title{Standardise possessives in address entry(/ies)}
\usage{
clean_address_possessives(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with clean possessive(s).
}
\description{
Attempts to standardise possessives in provided address entry(/ies).
}
