% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_pre_clean}
\alias{clean_address_pre_clean}
\title{Pre-cleaning operation for address entry(/ies)}
\usage{
clean_address_pre_clean(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector with address(es) cleaner than the one
provided in \code{addresses}.
}
\description{
Performs pre-cleaning operations on provided address entry(/ies).
}
