% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_saints}
\alias{clean_address_saints}
\title{Clean "Saint" prefix in address entry(/ies)}
\usage{
clean_address_saints(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with clean "Saint" prefix(es).
}
\description{
Attempts to clean "Saint" prefix in provided address entry(/ies).
}
