% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_address_suffixes}
\alias{clean_address_suffixes}
\title{Clean unwanted suffixes in address entry(/ies)}
\usage{
clean_address_suffixes(addresses)
}
\arguments{
\item{addresses}{A character string vector of address(es).}
}
\value{
A character string vector of address(es) with unwanted suffix(es) removed.
}
\description{
Attempts to clean unwanted suffixes in provided address entry(/ies).
}
