% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_punctuation}
\alias{clean_forename_punctuation}
\title{Standardise punctuation in forename(s)}
\usage{
clean_forename_punctuation(forenames)
}
\arguments{
\item{forenames}{A character string vector of forename(s).}
}
\value{
A character string vector of forename(s) with clean punctuation.
}
\description{
Attempts to standardise punctuation in provided forename entry(/ies).
}
