% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_separate_words}
\alias{clean_forename_separate_words}
\title{Separate double-barrelled forename(s)}
\usage{
clean_forename_separate_words(forenames)
}
\arguments{
\item{forenames}{A character string vector of forename(s).}
}
\value{
A character string vector of forename(s) with clean double-barrelled
forename(s).
}
\description{
Attempts to separate double-barrelled forename(s) in provided forename entry(/ies).
}
