% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_forename_spelling}
\alias{clean_forename_spelling}
\title{Clean forename(s) spelling}
\usage{
clean_forename_spelling(forenames)
}
\arguments{
\item{forenames}{A character string vector of forename(s).}
}
\value{
A character string vector of forename(s) with clean forename(s) spelling.
}
\description{
Attempts to clean spelling in provided forename entry(/ies).
}
