% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_mac}
\alias{clean_mac}
\title{Standardise "Mac" prefix in people's name}
\usage{
clean_mac(names)
}
\arguments{
\item{names}{A character string vector of name(s).}
}
\value{
A character string vector of name(s) with clean "Mac" prefix(es).
}
\description{
Attempts to standardise "Mac" prefix in provided name entry(/ies).
}
