% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_name_ends}
\alias{clean_name_ends}
\title{Clean ends in entry(/ies) names}
\usage{
clean_name_ends(names)
}
\arguments{
\item{names}{A character string vector of names}
}
\value{
A character string vector of names with clean ends.
}
\description{
Attempts to clean ends in provided name entry(/ies).
}
