% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_occupation}
\alias{clean_occupation}
\title{Clean entry(/ies) occupation}
\usage{
clean_occupation(occupations)
}
\arguments{
\item{occupations}{A character string vector of occupation(s).}
}
\value{
A character string vector of cleaned occupation(s).
}
\description{
Attempts to clean provided occupation.
}
