% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_parentheses}
\alias{clean_parentheses}
\title{Clean entry(/ies) of in brackets information}
\usage{
clean_parentheses(x)
}
\arguments{
\item{x}{A character string vector.}
}
\value{
A character string vector with within brackets content removed.
}
\description{
Attempts to clean entry(/ies) of unwanted information displayed in brackets.
}
