% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_specials}
\alias{clean_specials}
\title{Clean entries special characters}
\usage{
clean_specials(x)
}
\arguments{
\item{names}{A character string vector of occupation(s).}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean entries of unwanted special characters
}
\examples{
clean_specials(c("Wine and spirit »merchant", "Mac» William"))
}
