% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_specials}
\alias{clean_specials}
\title{Clean entry(/ies) special characters}
\usage{
clean_specials(x)
}
\arguments{
\item{x}{A character string vector.}
}
\value{
A character string vector with special character(s) removed.
}
\description{
Attempts to clean entry(/ies) of unwanted special character(s).
}
