% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_string_ends}
\alias{clean_string_ends}
\title{Clean string ends}
\usage{
clean_string_ends(strings)
}
\arguments{
\item{strings}{A character string vector.}
}
\value{
A character string vector with clean entry(/ies) ends.
}
\description{
Attempts to clean ends of strings provided.
}
