% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_surname_punctuation}
\alias{clean_surname_punctuation}
\title{Standardise punctuation in surname(s)}
\usage{
clean_surname_punctuation(surnames)
}
\arguments{
\item{surnames}{A character string vector of surname(s).}
}
\value{
A character string vector of surname(s) with clean punctuation.
}
\description{
Attempts to standardise punctuation in provided surname entry(/ies).
}
