% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_surname_spelling}
\alias{clean_surname_spelling}
\title{Clean surnames spelling}
\usage{
clean_surname_spelling(name)
}
\arguments{
\item{names}{A character string vector of names.}
}
\value{
A vector of character strings.
}
\description{
Attempts to clean spelling in provided name entries.
}
\examples{
clean_surname_spelling(c("ABOT junior", "ABRCROMBIE", "BLAI senior"))
}
