% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_surname_spelling}
\alias{clean_surname_spelling}
\title{Clean surname(s) spelling}
\usage{
clean_surname_spelling(surnames)
}
\arguments{
\item{surnames}{A character string vector of surnames.}
}
\value{
A character string vector of surnames with clean spelling.
}
\description{
Attempts to clean spelling in provided surname entry(/ies).
}
