% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.r
\name{clean_title}
\alias{clean_title}
\title{Clean entry(/ies) name title}
\usage{
clean_title(names)
}
\arguments{
\item{names}{A character string vector of name(s).}
}
\value{
A character string vector of name(s) with cleaned title(s).
}
\description{
Attempts to clean titles attached to names provided: Captain, Major, etc.
}
