% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_has_match_failed}
\alias{combine_has_match_failed}
\title{Check for failed matches}
\usage{
combine_has_match_failed(number, body)
}
\arguments{
\item{number}{A vector of address number(s). Integer or character string.}

\item{body}{A character string vector of address body(/ies).}
}
\value{
A boolean vector: \code{TRUE} for indexes where both
\code{number} and \code{body} are "NA", \code{FALSE} otherwise.
}
\description{
Provided with two equal length vectors, returns TRUE for indexes where both
entries are "NA" and FALSE otherwise.
}
