% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_label_if_match_failed}
\alias{combine_label_if_match_failed}
\title{Label failed matches}
\usage{
combine_label_if_match_failed(type = c("number", "body"), ...)
}
\arguments{
\item{type}{A Character string, one of: "number" or "body". Type of column to
label.}

\item{...}{Further arguments to be passed down to
\code{\link{combine_has_match_failed}}}
}
\value{
A character string vector: address(es) "number" or "body" as
specified in \code{type} if match succeeded, "" (type = "number") or
"Failed to match with general directory" (type = "body") otherwise.
}
\description{
Labels failed matches as such.
}
