% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine_random_string_if_no_address}
\alias{combine_random_string_if_no_address}
\title{Conditionally return a random string}
\usage{
combine_random_string_if_no_address(address)
}
\arguments{
\item{address}{A character string.}
}
\value{
A length 1 character string vector: 22 character long random string
if \code{address} labelled as missing ("No trade/house address found"),
\code{address} otherwise.
}
\description{
Returns a 22 character long random string if address provided is labelled as
missing ("No trade/house address found").
}
