% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trades.r
\name{trades_clean_entries}
\alias{trades_clean_entries}
\title{Mutate operation(s) in trades directory dataframe column(s)}
\usage{
trades_clean_entries(directory, verbose)
}
\arguments{
\item{directory}{A trades directory dataframe. Columns must include \code{occupation},
\code{forename}, \code{surname}, \code{address.trade.number}, \code{address.trade.body}.}

\item{verbose}{Whether the function should be executed silently (\code{FALSE}) or
not (\code{TRUE}).}
}
\value{
A dataframe.
}
\description{
Attempts to clean entries of the provided trades directory dataframe provided.
}
\examples{
directory <- data.frame(
  page = c("71", "71"),
  surname = c("ABOT", "ABRCROMBIE"), forename = c("Wm.", "Alex"),
  occupation = c("Wine and spirit mercht — See Advertisement in Appendix.", "Bkr"),
  address.trade.number = c("1S20", "I2"),
  address.trade.body = c("Londn rd.", "Dixen pl"),
  stringsAsFactors = FALSE
)
trades_clean_entries(directory, verbose = FALSE)
}
